General improvements and others
Added the DialogBox class to handle exceptions that occur in the main window thread.
Added methods to add single-trigger listeners to any other ordered event.

Updated the Write and Writel commands to receive text in any case and handle quotes. They can now also accept variable names to output their values.
Improved the If and Batch commands to accept arguments in a different format.

Destroy in ModularObject and Dispose in ObjectModule now are called in EndSingle event. The virtual Post- and Pre- methods for the Dispose and Destroy have been restored.