Huge changes and improvements:
Added:
The Key enum - represents the Keys enum from XNA but includes mouse buttons. Input now works with it.
The WriteState and WriteError methods in the Monoconsole - allow writing with specialized colors.
ObjectExtensions class - contains extensions for any object, such as type casting.
ILevelObject - an interface that allows an object to be included in the list of objects in the Level class.
WhereX, WhereY, Where methods in Vector2Extensions - create another Vector with different X, Y, or the same values.

Added some commands to the Executor:
fps - sets the target FPS.
isfpsfixed - enables or disables targeting FPS that can be set using the fps command.
writeinput - shows the name of the last pressed key in the UI.
captwin - allows creating screenshots of the game.
stopwatch - allows counting execution time.
writeexec - enables or disables writing commands that were executed (such as "> ram" in the Initialize).
batchfile - allows executing commands from a text file (such as batchbegin and batchend).
on - queues executing a command at a specialized event in the game (such as Update, EndUpdate, etc.).

Changed:
Level - added a list of objects that can be accessed using a tag.
Destroy in ModularObject - now forces Dispose methods of removed modules, allowing destruction of an object in one frame.