Improvements and game development
Added:
Thread safety for the LockCollection (lock).
Throwing an exception in EnumerableExtensions.RandomElement if the collection is empty.
The DrawOptions for the Drawer.
The PointTouchTask - a level task that gives a combo every few touches of a point, which moves to a new position every time the player touches it.
The LevelGenerator - totally remade the way to create levels; now it uses pictures and places objects depending on the color of each pixel.
The IPersistentObject - an object that cannot be removed from the level object list automatically when a level has been recreated/loaded.