The Final Update (probably)
Over the past 7 months, I've dedicated a lot of effort to this project. However, I've realized that both the idea and its implementation aren't as solid as I hoped. This is likely the final update, but elements of this idea might find new life in my future projects, albeit in a different form.